seeds are now fixed to be of type uint32_t

+++ base/EvolutionaryAlgorithm.h
@@ -33,11 +33,11 @@ namespace ea {
         /**
          * current seed for the RNG
          */
-        long seed;
+        uint32_t seed;
         /**
          * base seed for each batch of runs
          */
-        long baseSeed;
+        uint32_t baseSeed;
         /**
          * statistics of the EA
          */
